# internal/swagger/v3/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "contact" : { },
    "description" : "A lightweight, flexible, elegant and full-featured RBAC scaffolding based on GIN + GORM 2.0 + Casbin 2.0 + Wire DI.",
    "title" : "gin-admin",
    "version" : "v10.1.0"
  },
  "servers" : [ {
    "url" : "/"
  } ],
  "paths" : {
...
    "/api/v1/users" : {
      "get" : {
        "parameters" : [ {
          "description" : "pagination index",
          "in" : "query",
          "name" : "current",
          "required" : true,
          "schema" : {
            "default" : 1,
            "type" : "integer"
          }
        }, {
          "description" : "pagination size",
          "in" : "query",
          "name" : "pageSize",
          "required" : true,
          "schema" : {
            "default" : 10,
            "type" : "integer"
          }
        }, {
          "description" : "Username for login",
          "in" : "query",
          "name" : "username",
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Name of user",
          "in" : "query",
          "name" : "name",
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Status of user (activated, freezed)",
          "in" : "query",
          "name" : "status",
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/_api_v1_users_get_200_response"
                }
              }
            },
            "description" : "OK"
          },
          "401" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/util.ResponseResult"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "500" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/util.ResponseResult"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "security" : [ {
          "ApiKeyAuth" : [ ]
        } ],
        "summary" : "Query user list",
        "tags" : [ "UserAPI" ]
      },
      "post" : {
        "requestBody" : {
          "content" : {
            "*/*" : {
              "schema" : {
                "$ref" : "#/components/schemas/schema.UserForm"
              }
            }
          },
          "description" : "Request body",
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/_api_v1_current_user_get_200_response"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/util.ResponseResult"
                }
              }
            },
            "description" : "Bad Request"
          },
          "401" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/util.ResponseResult"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "500" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/util.ResponseResult"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "security" : [ {
          "ApiKeyAuth" : [ ]
        } ],
        "summary" : "Create user record",
        "tags" : [ "UserAPI" ],
        "x-codegen-request-body-name" : "body"
      }
    },
    "/api/v1/users/{id}" : {
      "delete" : {
        "parameters" : [ {
          "description" : "unique id",
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/util.ResponseResult"
                }
              }
            },
            "description" : "OK"
          },
          "401" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/util.ResponseResult"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "500" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/util.ResponseResult"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "security" : [ {
          "ApiKeyAuth" : [ ]
        } ],
        "summary" : "Delete user record by ID",
        "tags" : [ "UserAPI" ]
      },
      "get" : {
        "parameters" : [ {
          "description" : "unique id",
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/_api_v1_current_user_get_200_response"
                }
              }
            },
            "description" : "OK"
          },
          "401" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/util.ResponseResult"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "500" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/util.ResponseResult"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "security" : [ {
          "ApiKeyAuth" : [ ]
        } ],
        "summary" : "Get user record by ID",
        "tags" : [ "UserAPI" ]
      },
      "put" : {
        "parameters" : [ {
          "description" : "unique id",
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "*/*" : {
              "schema" : {
                "$ref" : "#/components/schemas/schema.UserForm"
              }
            }
          },
          "description" : "Request body",
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/util.ResponseResult"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/util.ResponseResult"
                }
              }
            },
            "description" : "Bad Request"
          },
          "401" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/util.ResponseResult"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "500" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/util.ResponseResult"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "security" : [ {
          "ApiKeyAuth" : [ ]
        } ],
        "summary" : "Update user record by ID",
        "tags" : [ "UserAPI" ],
        "x-codegen-request-body-name" : "body"
      }
    },
    "/api/v1/users/{id}/reset-pwd" : {
      "patch" : {
        "parameters" : [ {
          "description" : "unique id",
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/util.ResponseResult"
                }
              }
            },
            "description" : "OK"
          },
          "401" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/util.ResponseResult"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "500" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/util.ResponseResult"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "security" : [ {
          "ApiKeyAuth" : [ ]
        } ],
        "summary" : "Reset user password by ID",
        "tags" : [ "UserAPI" ]
      }
    }
  },
  "components" : {
    "schemas" : {
      "errors.Error" : {
        "properties" : {
          "code" : {
            "type" : "integer"
          },
          "detail" : {
            "type" : "string"
          },
          "id" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "schema.User" : {
        "properties" : {
          "created_at" : {
            "description" : "Create time",
            "type" : "string"
          },
          "email" : {
            "description" : "Email of user",
            "type" : "string"
          },
          "id" : {
            "description" : "Unique ID",
            "type" : "string"
          },
          "name" : {
            "description" : "Name of user",
            "type" : "string"
          },
          "phone" : {
            "description" : "Phone number of user",
            "type" : "string"
          },
          "remark" : {
            "description" : "Remark of user",
            "type" : "string"
          },
          "roles" : {
            "description" : "Roles of user",
            "items" : {
              "$ref" : "#/components/schemas/schema.UserRole"
            },
            "type" : "array"
          },
          "status" : {
            "description" : "Status of user (activated, freezed)",
            "type" : "string"
          },
          "updated_at" : {
            "description" : "Update time",
            "type" : "string"
          },
          "username" : {
            "description" : "Username for login",
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "schema.UserForm" : {
        "properties" : {
          "email" : {
            "description" : "Email of user",
            "maxLength" : 128,
            "type" : "string"
          },
          "name" : {
            "description" : "Name of user",
            "maxLength" : 64,
            "type" : "string"
          },
          "password" : {
            "description" : "Password for login (md5 hash)",
            "maxLength" : 64,
            "type" : "string"
          },
          "phone" : {
            "description" : "Phone number of user",
            "maxLength" : 32,
            "type" : "string"
          },
          "remark" : {
            "description" : "Remark of user",
            "maxLength" : 1024,
            "type" : "string"
          },
          "roles" : {
            "description" : "Roles of user",
            "items" : {
              "$ref" : "#/components/schemas/schema.UserRole"
            },
            "type" : "array"
          },
          "status" : {
            "description" : "Status of user (activated, freezed)",
            "enum" : [ "activated", "freezed" ],
            "type" : "string"
          },
          "username" : {
            "description" : "Username for login",
            "maxLength" : 64,
            "type" : "string"
          }
        },
        "required" : [ "name", "roles", "status", "username" ],
        "type" : "object"
      },
      "schema.UserRole" : {
        "properties" : {
          "created_at" : {
            "description" : "Create time",
            "type" : "string"
          },
          "id" : {
            "description" : "Unique ID",
            "type" : "string"
          },
          "role_id" : {
            "description" : "From Role.ID",
            "type" : "string"
          },
          "role_name" : {
            "description" : "From Role.Name",
            "type" : "string"
          },
          "updated_at" : {
            "description" : "Update time",
            "type" : "string"
          },
          "user_id" : {
            "description" : "From User.ID",
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "util.ResponseResult" : {
        "properties" : {
          "data" : {
            "type" : "object"
          },
          "error" : {
            "$ref" : "#/components/schemas/errors.Error"
          },
          "success" : {
            "type" : "boolean"
          },
          "total" : {
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "_api_v1_current_user_get_200_response" : {
        "allOf" : [ {
          "$ref" : "#/components/schemas/util.ResponseResult"
        }, {
          "properties" : {
            "data" : {
              "$ref" : "#/components/schemas/schema.User"
            }
          },
          "type" : "object"
        } ]
      },
...
      "_api_v1_users_get_200_response" : {
        "allOf" : [ {
          "$ref" : "#/components/schemas/util.ResponseResult"
        }, {
          "properties" : {
            "data" : {
              "items" : {
                "$ref" : "#/components/schemas/schema.User"
              },
              "type" : "array"
            }
          },
          "type" : "object"
        } ]
      }
    },
    "securitySchemes" : {
      "ApiKeyAuth" : {
        "in" : "header",
        "name" : "Authorization",
        "type" : "apiKey"
      }
    }
  },
...
}